[ROCm] Adding pass to lower GPU Dialect to ROCDL Dialect.
This is a follow-up to the PR#146 which introduced the ROCDL Dialect. This PR introduces a pass to lower GPU Dialect to the ROCDL Dialect. As with the previous PR, this one builds on the work done by @whchung, and addresses most of the review comments in the original PR.