Add uitofp to the standard dialect.
The operation 'uitofp', which converts an unsigned integer to a floating-point
value was already part of the LLVM dialect. Contrary to its sibling 'sitofp' it
had no corresponding operation in the standard dialect.

This adds UIToFPOp to the standard dialect and a lowering to 'uitofp' in LLVM.
The work is mostly copy-paste of the existing SIToFPOp.